+++ epubreader.php
@@ -12,7 +12,4 @@ use SebLucas\Cops\Framework;
 
 require_once __DIR__ . '/config.php';
 
-$request = Framework::getRequest('read');
-
-$handler = Framework::getHandler('read');
-$handler->handle($request);
+Framework::run('read');