+++ front.php
@@ -16,6 +16,10 @@ $request = Framework::getRequest();
 
 // @todo route to the right handler if needed
 $name = $request->getHandler();
+// special case for json requests here
+if ($name == 'index' && $request->isJson()) {
+    $name = 'json';
+}
 $handler = Framework::getHandler($name);
 //$handler->handle($request);
 var_dump($_SERVER);