+++ getJSON.php
@@ -13,8 +13,4 @@ use SebLucas\Cops\Framework;
 require_once __DIR__ . '/config.php';
 
 // @todo handle 'json' routes correctly - see util.js
-//$request = Framework::getRequest('json');
-$request = Framework::getRequest('index');
-
-$handler = Framework::getHandler('json');
-$handler->handle($request);
+Framework::run('json');