+++ sendtomail.php
@@ -12,7 +12,4 @@ use SebLucas\Cops\Framework;
 
 require_once __DIR__ . '/config.php';
 
-$request = Framework::getRequest('mail');
-
-$handler = Framework::getHandler('mail');
-$handler->handle($request);
+Framework::run('mail');